+++ components/platform/hosal/include/hosal_spi.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2022 Bouffalolab.
+ * Copyright (c) 2016-2023 Bouffalolab.
  *
  * This file is part of
  *     *** Bouffalolab Software Dev Kit ***
@@ -98,7 +98,7 @@ int hosal_spi_init(hosal_spi_dev_t *spi);
  *       - 0 : on success  
  *       - other : error
  */
-int hosal_spi_send(hosal_spi_dev_t *spi, const uint8_t *data, uint32_t size, uint32_t timeout);
+int hosal_spi_send(hosal_spi_dev_t *spi, const uint8_t *data, uint16_t size, uint32_t timeout);
 
 /**
  * @brief Spi recv
